+++ src/pages/home.php
@@ -1,0 +1,30 @@
+<!-- @TODO: Implement recently viewed products -->
+    <section title="Recently Viewed posts">
+        <h2>Recently Viewed Posts</h2>
+        <div class="post-list">
+            <div class="post">
+                <a href="/post-1-url" title="Post 1">
+                    <img src="/post-placeholder.png" alt="Post 1" width="200"/>
+                </a>
+                <p><a href="/post-1-url" title="Post 1">Post 1</a></p>
+                <span>Oleksandr Babii</span>
+                <span>2021-12-05</span>
+            </div>
+            <div class="post">
+                <a href="/post-2-url" title="Post 2">
+                    <img src="/post-placeholder.png" alt="Post 2" width="200"/>
+                </a>
+                <p><a href="/post-2-url" title="Post 2">Post 2</a></p>
+                <span>Vitaliy Kayun</span>
+                <span>2021-12-15</span>
+            </div>
+            <div class="post">
+                <a href="/post-3-url" title="Post 3">
+                    <img src="/post-placeholder.png" alt="Post 3" width="200"/>
+                </a>
+                <p><a href="/post-3-url" title="Post 3">Post 3</a></p>
+                <span>Oleksandr Babii</span>
+                <span>2021-12-09</span>
+            </div>
+        </div>
+    </section>
