+++ src/pages/post.php
@@ -1,0 +1,7 @@
+<div class="post-list">
+    <img src="/post-placeholder.png" alt="<?= $data['name'] ?>" width="300"/>
+    <h1><?= $data['name'] ?></h1>
+    <p><?= $data['description'] ?></p>
+    <span><?= $data['author'] ?></span>
+    <span><?= $data['date'] ?></span>
+</div>
